feat: Introduce EntityType enum and enhance service and venue handling with new queries and command updates

+++ Application/Features/User/Commands/UpdateUserRecord/UpdateUserRecordCommandHandler.cs
@@ -74,7 +74,7 @@ public sealed class UpdateUserRecordCommandHandler(
 
         if (record.Status == RecordStatus.Completed && request.Review is not null)
         {
-            result = record.SetReview((byte)request.Review.Rating, request.Review.Comment);
+            result = record.SetReview(request.Review.Rating, request.Review.Comment);
             if (result.IsFailure)
                 return result;
